Use template for snap account creation
Move CreateSnapAccount component out of ConfirmationPage.
Pass account name to onSubmit in template.
Remove buttons from CreateSnapAccount.
