+++ ui/pages/confirmation/templates/index.js
@@ -6,7 +6,7 @@ import {
   setNewNetworkAdded,
   upsertNetworkConfiguration,
 } from '../../../store/actions';
-import createSnapAccountTemplate from './create-snap-account/create-snap-account-template';
+import createSnapAccount from './create-snap-account';
 import addEthereumChain from './add-ethereum-chain';
 import switchEthereumChain from './switch-ethereum-chain';
 import success from './success';
@@ -27,7 +27,7 @@ const APPROVAL_TEMPLATES = {
   [ApprovalType.SnapDialogAlert]: snapAlert,
   [ApprovalType.SnapDialogConfirmation]: snapConfirmation,
   [ApprovalType.SnapDialogPrompt]: snapPrompt,
-  'snap_manageAccounts:confirmation': createSnapAccountTemplate,
+  'snap_manageAccounts:confirmation': createSnapAccount,
   ///: END:ONLY_INCLUDE_IN
 };
