+++ ui/pages/create-snap-account/components/index.ts
