+++ ui/pages/create-snap-account/index.scss
