+++ ui/pages/create-snap-account/index.ts
